.history/requirements_20240509172705.txt
colorama==0.4.6
crayons==0.4.0
gnupg==2.3.1
numpy==1.26.4
opencv-python==4.9.0.80
piexif==1.1.3
pillow==10.3.0
psutil==5.9.8
python-gnupg==0.5.2
stegano==0.11.3
